refactor(app): removed console errors
The console errors being shown included:
- unique keys not being used in map functions
-console.log messages showing for debugging purposes
- passing undefined values to GatsbyImage componenet

Gatsby Image componenet remains a challenge. Gatsby-image-plugin
does not seem to allow for a blank soure image.src='' on the first node
within a JSON file. However, a blank source on subsequent nodes seems to
be allowed. Not sure why this is happening. My solution to remove the
image.src from the 1st node of the structured-resume.json file is a very
hacky way of fixing this issue. If we were delivered an API with an
existing blank source i.e. image.src = "" this would break the app